requirements/local.txt
-r base.txt
django-debug-toolbar==3.2.2
gunicorn==20.1.0
whitenoise==5.3.0
django-mptt==0.13.4